+++ dashboard/lib/flanker/addresslib/_parser/parser.py
@@ -1,12 +1,11 @@
-import ply.yacc as yacc
-from lexer import tokens, lexer
-from collections import namedtuple
 import logging
+from collections import namedtuple
 
+import ply.yacc as yacc
 
-logging.basicConfig()
+from flanker.addresslib._parser.lexer import lexer, tokens
+
 log = logging.getLogger(__name__)
-log.setLevel(logging.INFO)
 
 Mailbox = namedtuple('Mailbox', ['display_name', 'local_part', 'domain'])
 Url     = namedtuple('Url',     ['address'])
@@ -156,26 +155,40 @@ def p_error(p):
 
 
 # Build the parsers
+log.debug('building mailbox parser')
+mailbox_parser = yacc.yacc(start='mailbox',
+                           errorlog=log,
+                           tabmodule='mailbox_parsetab',
+                           debug=False,
+                           write_tables=False)
 
-log.info('building mailbox parser')
-mailbox_parser = yacc.yacc(
-    start='mailbox', errorlog=log)
+log.debug('building addr_spec parser')
+addr_spec_parser = yacc.yacc(start='addr_spec',
+                             errorlog=log,
+                             tabmodule='addr_spec_parsetab',
+                             debug=False,
+                             write_tables=False)
 
-log.info('building addr_spec parser')
-addr_spec_parser = yacc.yacc(
-    start='addr_spec', errorlog=log)
+log.debug('building url parser')
+url_parser = yacc.yacc(start='url',
+                       errorlog=log,
+                       tabmodule='url_parsetab',
+                       debug=False,
+                       write_tables=False)
 
-log.info('building url parser')
-url_parser = yacc.yacc(
-    start='url', errorlog=log)
+log.debug('building mailbox_or_url parser')
+mailbox_or_url_parser = yacc.yacc(start='mailbox_or_url',
+                                  errorlog=log,
+                                  tabmodule='mailbox_or_url_parsetab',
+                                  debug=False,
+                                  write_tables=False)
 
-log.info('building mailbox_or_url parser')
-mailbox_or_url_parser = yacc.yacc(
-    start='mailbox_or_url', errorlog=log)
-
-log.info('building mailbox_or_url_list parser')
-mailbox_or_url_list_parser = yacc.yacc(
-    start='mailbox_or_url_list', errorlog=log)
+log.debug('building mailbox_or_url_list parser')
+mailbox_or_url_list_parser = yacc.yacc(start='mailbox_or_url_list',
+                                       errorlog=log,
+                                       tabmodule='mailbox_or_url_list_parsetab',
+                                       debug=False,
+                                       write_tables=False)
 
 
 # Interactive prompt for easy debugging
@@ -189,16 +202,16 @@ if __name__ == '__main__':
             break
         if s == '': continue
 
-        print '\nTokens list:\n'
+        print('\nTokens list:\n')
         lexer.input(s)
         while True:
             tok = lexer.token()
             if not tok:
                 break
-            print tok
+            print(tok)
 
-        print '\nParsing behavior:\n'
+        print('\nParsing behavior:\n')
         result = mailbox_or_url_list_parser.parse(s, debug=log)
 
-        print '\nResult:\n'
-        print result
+        print('\nResult:\n')
+        print(result)